fix on 32-bit

+++ rpython/rlib/rvmprof/src/rvmprof.h
@@ -28,7 +28,7 @@ RPY_EXTERN char *vmprof_init(int fd, double interval, int memory,
 RPY_EXTERN void vmprof_ignore_signals(int);
 RPY_EXTERN int vmprof_enable(int memory, int native);
 RPY_EXTERN int vmprof_disable(void);
-RPY_EXTERN int vmprof_register_virtual_function(char *, long, int);
+RPY_EXTERN int vmprof_register_virtual_function(char *, intptr_t, int);
 RPY_EXTERN void* vmprof_stack_new(void);
 RPY_EXTERN int vmprof_stack_append(void*, long);
 RPY_EXTERN long vmprof_stack_pop(void*);
